fix unknown command (#246)
